+++ pandas/tests/io/excel/conftest.py
@@ -1,5 +1,7 @@
 import pytest
 
+import pandas.util._test_decorators as td
+
 import pandas.util.testing as tm
 
 from pandas.io.parsers import read_csv
@@ -39,3 +41,25 @@ def read_ext(request):
     Valid extensions for reading Excel files.
     """
     return request.param
+
+
+@pytest.fixture(autouse=True)
+def check_for_file_leaks():
+    """
+    Fixture to run around every test to ensure that we are not leaking files.
+
+    See also
+    --------
+    _test_decorators.check_file_leaks
+    """
+    # GH#30162
+    psutil = td.safe_import("psutil")
+    if not psutil:
+        yield
+
+    else:
+        proc = psutil.Process()
+        flist = proc.open_files()
+        yield
+        flist2 = proc.open_files()
+        assert flist == flist2
